Move weak references implementation to script_bindings.
